allow use as a module
I was importing setcola as a module: this worked fine when using the webpack dev server, but threw an error when I opened the site generated by "npm run build". Specifically, "this" is undefined at the top level of a module, so `setcola.nodes(nodes).links(links)` threw an error as `setcola.node()` returned `undefined`.